+++ lib/libvector/vector.c
@@ -6,13 +6,13 @@
 /*   By: aroque <aroque@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/14 16:01:16 by aroque            #+#    #+#             */
-/*   Updated: 2020/07/08 10:16:53 by aroque           ###   ########.fr       */
+/*   Updated: 2020/09/01 22:30:43 by aroque           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector.h"
 
-t_vector	vector(float x, float y, float z)
+t_vector	vector(double x, double y, double z)
 {
 	t_vector v;
 
@@ -22,7 +22,7 @@ t_vector	vector(float x, float y, float z)
 	return (v);
 }
 
-t_point		point(float x, float y, float z)
+t_point		point(double x, double y, double z)
 {
 	return (vector(x, y, z));
 }
