+++ lib/libvector/vector.h
@@ -6,7 +6,7 @@
 /*   By: aroque <aroque@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/14 15:56:30 by aroque            #+#    #+#             */
-/*   Updated: 2020/07/13 10:49:15 by aroque           ###   ########.fr       */
+/*   Updated: 2020/09/01 22:36:27 by aroque           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,22 +15,22 @@
 
 typedef struct		s_vector
 {
-	float			x;
-	float			y;
-	float			z;
+	double			x;
+	double			y;
+	double			z;
 }					t_vector;
 
 typedef	t_vector	t_point;
 
-t_vector			vector(float x, float y, float z);
-t_point				point(float x, float y, float z);
+t_vector			vector(double x, double y, double z);
+t_point				point(double x, double y, double z);
 t_vector			add(t_vector u, t_vector v);
 t_vector			sub(t_vector u, t_vector v);
-t_vector			scale(t_vector u, float c);
+t_vector			scale(t_vector u, double c);
 t_vector			cross(t_vector u, t_vector v);
-float				dot(t_vector u, t_vector v);
-float				length_squared(t_vector u);
-float				length(t_vector u);
+double				dot(t_vector u, t_vector v);
+double				length_squared(t_vector u);
+double				length(t_vector u);
 t_vector			norm(t_vector u);
 
 #endif
